More and More and More and More refactoring
