+++ ModCore/ContextMenu/MemberContextMenu.cs
@@ -1,0 +1,9 @@
+﻿using DSharpPlus.SlashCommands;
+
+namespace ModCore.ContextMenu
+{
+    public class MemberContextMenu : ApplicationCommandModule
+    {
+        // Member context menu commands here. Max 5.
+    }
+}
