+++ ModCore/ContextMenu/MessageContextMenu.cs
@@ -1,0 +1,9 @@
+﻿using DSharpPlus.SlashCommands;
+
+namespace ModCore.ContextMenu
+{
+    public class MessageContextMenu : ApplicationCommandModule
+    {
+        // Message Context Menu commands here. Max 5.
+    }
+}
