+++ ModCore/Database/DatabaseContext.cs
@@ -6,7 +6,7 @@ using System.Threading.Tasks;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using ModCore.Database.Entities;
+using ModCore.Database.DatabaseEntities;
 using ModCore.Entities;
 using ModCore.Logic;
 using ModCore.Logic.EntityFramework;
