+++ ModCore/Database/DatabaseEntities/DatabaseCommandId.cs
@@ -4,7 +4,7 @@ using Microsoft.EntityFrameworkCore.Metadata;
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database.Entities
+namespace ModCore.Database.DatabaseEntities
 {
     [Table("mcore_cmd_state")]
     public class DatabaseCommandId
