+++ ModCore/Database/DatabaseEntities/DatabaseGuildConfig.cs
@@ -1,10 +1,11 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
+using ModCore.Database.JsonEntities;
 using ModCore.Entities;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 using Newtonsoft.Json;
 
-namespace ModCore.Database.Entities
+namespace ModCore.Database.DatabaseEntities
 {
     [Table("mcore_guild_config")]
     public class DatabaseGuildConfig
