+++ ModCore/Database/DatabaseEntities/DatabaseModNote.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations.Schema;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database.Entities
+namespace ModCore.Database.DatabaseEntities
 {
     [Table("mcore_modnotes")]
     public class DatabaseModNote
