+++ ModCore/Database/DatabaseEntities/DatabaseRolestateNick.cs
@@ -3,7 +3,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using ModCore.Entities;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database.Entities
+namespace ModCore.Database.DatabaseEntities
 {
     [Table("mcore_rolestate_nicks")]
     public class DatabaseRolestateNick
