+++ ModCore/Database/DatabaseEntities/DatabaseRolestateOverride.cs
@@ -2,7 +2,7 @@
 using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using ModCore.Logic.EntityFramework.AttributeImpl;
 
-namespace ModCore.Database.Entities
+namespace ModCore.Database.DatabaseEntities
 {
     [Table("mcore_rolestate_overrides")]
     public class DatabaseRolestateOverride
