+++ ModCore/Database/JsonEntities/UserData.cs
@@ -3,7 +3,7 @@ using System;
 using System.Collections.Generic;
 using System.Text;
 
-namespace ModCore.Entities
+namespace ModCore.Database.JsonEntities
 {
     public class UserData
     {
