+++ ModCore/Entities/SharedData.cs
@@ -6,6 +6,7 @@ using System.Threading;
 using DSharpPlus;
 using DSharpPlus.CommandsNext;
 using DSharpPlus.Entities;
+using ModCore.Database.JsonEntities;
 using ModCore.Listeners;
 
 namespace ModCore.Entities
@@ -22,41 +23,11 @@ namespace ModCore.Entities
         public ConcurrentDictionary<ulong, DiscordMessage> DeletedMessages = new ConcurrentDictionary<ulong, DiscordMessage>();
         public ConcurrentDictionary<ulong, DiscordMessage> EditedMessages = new ConcurrentDictionary<ulong, DiscordMessage>();
 
-        /// <summary>
-        /// Every command, top-level or not, along with full qualified name.
-        /// </summary>
-        public (string name, Command cmd)[] Commands { get; set; }
-
-        public string ApiToken = null;
-
         public ModCore ModCore;
 
         public SharedData()
         {
             this.TimerSempahore = new SemaphoreSlim(1, 1);
-        }
-
-        public void Initialize(ModCoreShard shard)
-        {
-            Commands = shard.Commands.RegisteredCommands.SelectMany(SelectCommandsFromDict).Distinct().ToArray();
-        }
-
-        private static IEnumerable<(string name, Command cmd)> SelectCommandsFromDict(KeyValuePair<string, Command> c)
-            => CommandSelector(c.Value);
-
-        private static IEnumerable<(string name, Command cmd)> CommandSelector(Command command)
-        {
-            yield return (command.QualifiedName, command);
-            if (!(command is CommandGroup group)) yield break;
-            if (group.Children == null) yield break;
-
-            foreach (var icommand in group.Children)
-            {
-                foreach (var res in CommandSelector(icommand))
-                {
-                    yield return res;
-                }
-            }
         }
 	}
 }
