+++ ModCore/Listeners/Reactions.cs
@@ -6,7 +6,8 @@ using DSharpPlus.Entities;
 using DSharpPlus.EventArgs;
 using Humanizer;
 using ModCore.Database;
-using ModCore.Database.Entities;
+using ModCore.Database.DatabaseEntities;
+using ModCore.Database.JsonEntities;
 using ModCore.Entities;
 using ModCore.Extensions.AsyncListeners.Attributes;
 using ModCore.Extensions.AsyncListeners.Enums;
