+++ ModCore/Listeners/Timers.cs
@@ -9,7 +9,8 @@ using DSharpPlus.EventArgs;
 using Humanizer;
 using Microsoft.Extensions.Logging;
 using ModCore.Database;
-using ModCore.Database.Entities;
+using ModCore.Database.DatabaseEntities;
+using ModCore.Database.JsonEntities;
 using ModCore.Entities;
 using ModCore.Extensions.AsyncListeners.Attributes;
 using ModCore.Extensions.AsyncListeners.Enums;
