+++ ModCore/Listeners/UnbanTimerRemove.cs
@@ -2,6 +2,7 @@
 using DSharpPlus;
 using DSharpPlus.EventArgs;
 using ModCore.Database;
+using ModCore.Database.JsonEntities;
 using ModCore.Entities;
 using ModCore.Extensions.AsyncListeners.Attributes;
 using ModCore.Extensions.AsyncListeners.Enums;
